Add logic to determine default service for unique id
Previously, the first `runCommand` call performed with a new client
would be sent to a DXL "commands" service topic, falling back to a
call to the DXL "remote" service in the event of a
"service unavailable" failure. If a "commands" service is registered on
the fabric under a different ePO unique id than the one provided during
client construction, however, the initial `runCommand` call would fail
with no response (due to being unauthorized) - never failing over to
the fallback call to the DXL "remote" service.

In this commit, the id supplied to the client is matched to a
"commands" or "remote" service at construction time - before the first
`runCommand` call is made. This avoids the unauthorized call to the
"commands" service in the event that the supplied unique id matches a
registered "remote" service. In the event that the supplied unique id
does not match any registered service, an `Error` is delivered to the
callback provided at client construction.