Look for remote service before commands service
Previously, the client would attempt to use an ePO "commands" service if
both a "remote" and "commands" service were available on the fabric.
Since no commands are authorized by default, a user who has a "remote"
service on the fabric would encounter client failures after upgrading
the ePO server to DXL 5 extensions until authorization is configured for
the commnands service.

This commit changes the logic to look for and use a "remote" service
first if both a "remote" and "commands" service are available. This
allows for calls to the remote service to conitue to function properly
for backward compatibility on initial upgrade of the ePO server to DXL 5
extensions.