Move ePO service type detection to runCommand
Previously, users would need to provide a callback parameter to the
EpoClient constructor and wait for the callback to be invoked before
making any runCommand() calls. This was necessary because the ePO
service type and id were determined before the callback was invoked.

With the changes in this commit, users no longer need to provide a
callback parameter to the EpoClient constructor. Instead, the client
will determine the ePO service type and id as part of performing the
runCommand() call. Once the initial service type and id have been
determined via a runCommand() call, subsequent runCommand() calls will
reuse locally cached service type and id information - to avoid having
to make additional queries to the broker service registry before each
subsequent ePO remote command which is performed.

This commit also moves the callback parameter provided to the
runCommand() to a property named `responseCallback`, which can be set
on the `options` parameter. This allows the `responseCallback` to be
optional.